refactor(store): move useConversationTabs hook into useSidebar store
