--- frontend/hooks/useConversationTabs.ts
@@ -1,43 +1,0 @@
-'use client';
-
-import { useState, useEffect } from 'react';
-import { getUserConversationsList } from '@/lib/api/messages';
-
-import { useSidebarStore } from '@/store/useSidebarStore';
-
-export function useConversationTabs() {
-  const conversationSnippets = useSidebarStore((state) => state.conversationSnippets);
-
-  const isLoading = useSidebarStore((state) => state.isLoading);
-
-  const setSnippets = useSidebarStore((state) => state.setSnippets);
-
-  const setLoading = useSidebarStore((state) => state.setLoading);
-
-  const [error, setError] = useState<string | null>(null);
-
-  const getConversations = async () => {
-    setLoading(true);
-    try {
-      const data = await getUserConversationsList();
-
-      setSnippets(data.conversations);
-
-      return { error: null };
-    } catch (err: unknown) {
-      const errorMessage = err instanceof Error ? err.message : 'An unexpected error occurred';
-
-      setError(errorMessage);
-
-      return { error: errorMessage };
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    getConversations();
-  }, []);
-
-  return { conversationSnippets, isLoading, error };
-}
